+++ CHANGELOG.md
@@ -7,6 +7,16 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [Unreleased] - yyyy-mm-dd
 
+## [1.28.0b1] - TBD
+
+### Added
+
+- `testing.reset_celery_once_locks`: Function for deleting celery once locks
+
+### Changed
+
+- `esi.retry_task_on_esi_error_and_offline`: Will now also retry when the rate limit is exceeded
+
 ## [1.27.0] - 2025-10-12
 
 ### Added